+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [2.6.4] - 2022-7-??
+ 
+This update exists as the first revisit of the codebase in over a year. With a fresh set of eyes, things are slowly being cleaned up and functionality of slash commands is taken with a fresh approach. discord.js v14 is introduced in this updated with v10 of the discord API, and most of the work in this iteration is to get the code up to speed to ensure it works with discord.js v14.
+
+### Added
+
+- Extra error checking is added for elo commands in case of invalid inputs or unexpected resutls [elo.js]
+- getScore command added to obtain a single score from a database [db_helper.js]
+ 
+### Changed
+
+- Changed all references to "elo" in the database to "score" for clarity. This includes function names and database references [db_helper.js, reroll.js, elo.js, match.js, setup.js]
+- Database checks abstracted out of elo function and into db_helper to better organize file priorities [elo.js, db_helper.js]
+ 
+### Fixed
+
+- The elo command changed to work with the native interaction type instead of message and args [elo.js]
  
 ## [2.6.3] - 2022-7-22